page_store: add lru cache and access hint
