+++ photondb/Cargo.toml
@@ -24,6 +24,7 @@ snap = "1.1.0"
 zstd = "0.12"
 zstd-safe = "6.0"
 crc32fast = "1.3"
+parking_lot = "0.12"
 
 [dev-dependencies]
 env_logger = "0.9.1"
